remove print statement, move repeated tests in wrong spot, save border weight mask correctly
